Add Condition Response

+++ rest-api/src/main/java/com/domain/controllers/PersonController.java
@@ -1,8 +1,10 @@
 package com.domain.controllers;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.integration.IntegrationProperties.Error;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,8 +24,6 @@ import com.domain.dto.Response;
 import com.domain.dto.SearchData;
 import com.domain.models.entities.Person;
 import com.domain.services.PersonService;
-
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 
 @RestController
@@ -36,10 +36,9 @@ public class PersonController {
 
   @PostMapping
   public ResponseEntity<Response<Person>> create(@Valid @RequestBody Person person, Errors errors) {
-
     Response<Person> responseData = new Response<>();
     HttpHeaders responseHeaders = new HttpHeaders();
-        responseHeaders.set("Content-type", "application/json");
+    responseHeaders.set("Content-type", "application/json");
 
     if(errors.hasErrors()) {
       for(ObjectError error : errors.getAllErrors()) {
@@ -49,6 +48,12 @@ public class PersonController {
       responseData.setPayload(null);
       return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseData);
     }
+    if(personService.checkingDataPerson(person.getNik())) {
+      responseData.setStatus(HttpStatus.BAD_REQUEST.value());
+      responseData.setPayload(null);
+      responseData.setMessages(List.of("NIK is already created"));
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseData);
+    }
     responseData.setStatus(HttpStatus.OK.value());
     responseData.setPayload(personService.save(person));
     return ResponseEntity.ok().headers(responseHeaders).body(responseData);
@@ -56,19 +61,20 @@ public class PersonController {
 
   @GetMapping
   public ResponseEntity<Iterable<Person>> findAll(){;
+    Response<Person> responseData = new Response<>();
+
+    responseData.setStatus(HttpStatus.OK.value());
     return ResponseEntity.ok(personService.findAll());
   }
 
   @GetMapping("/{nik}")
-  public ResponseEntity<Response<Person>> findOne(@Valid @PathVariable("nik") Long nik, Errors errors) {
+  public ResponseEntity<Response<Person>> findOne(@PathVariable("nik") Long nik) {
     Response<Person> responseData = new Response<>();
 
-    if(errors.hasErrors()) {
-      for(ObjectError error : errors.getAllErrors()) {
-        responseData.getMessages().add(error.getDefaultMessage());
-      }
+    if(!personService.checkingDataPerson(nik)) {
+      responseData.setMessages(List.of("NIK is Not found"));
       responseData.setStatus(HttpStatus.BAD_REQUEST.value());
-      responseData.setPayload(null);
+      responseData.setPayload(personService.findOne(nik));
       return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseData);
     }
     responseData.setStatus(HttpStatus.OK.value());
@@ -111,8 +117,19 @@ public class PersonController {
   }
 
   @DeleteMapping("/{nik}")
-  public void delete(@PathVariable("nik") Long nik) {
+  public ResponseEntity<Response<Person>> delete(@PathVariable("nik") Long nik) {
+    Response<Person> responseData = new Response<>();
+
+    if(!personService.checkingDataPerson(nik)) {
+      responseData.setMessages(List.of("Id not found"));
+      responseData.setStatus(HttpStatus.BAD_REQUEST.value());
+      responseData.setPayload(null);
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseData);
+    }
+    responseData.setStatus(HttpStatus.OK.value());
+    responseData.setPayload(null);
     personService.removeOne(nik);
+    return ResponseEntity.ok(responseData);
   }
 
 }
